deribit_positions/requirements.txt
python-deribit-api==1.2.0
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
plotly==5.18.0
python-dotenv==1.0.0
requests==2.31.0